😢  FFFFFaaaaaiiiiilllll
I keep on getting cannot read protocol of undefined error.

I think this is because one node_module is trying this

global.location.protocol

And global is something like service worker. I guess it is service worker because that's how Cloudflare workers work.

So I'll pause developing this for now.